requirements.txt
qdrant-client==1.7.0
cohere==4.37
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
